refactor: extract 5 sub-components from dashboard page (685→477 lines)
Move FlujoWaterfall, FlujoCharts, PresupuestoSection, PatrimonioSection,
and ActividadHeatmap into components/dashboard/. Remove unused
WidgetPlaceholder and now-unneeded imports from the route file.
